Adds can-related data structures to new can namespace
