+++ src/CanBehavior.cpp
@@ -1,5 +1,6 @@
 #include "CanBehavior.h"
 
-void CanBehavior::setSender(Sender* sender) {
+
+void CanBehavior::setSender(Delegate* sender) {
 	_sender = sender;
 }
